+++ Assets/Scripts/Event/EffectEvents.cs
@@ -1,8 +1,33 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class ParticleEvent : Event
+public abstract class EffectEvent : Event
+{
+    
+}
+
+[Serializable]
+public class HurtEffectEvent : EffectEvent
+{
+    public float amount;
+    public override IEnumerator Execute(EventObject eventObject = null)
+    {
+        return EffectManager.Instance.HurtEffect(amount);
+    }
+}
+
+[Serializable]
+public class ShakeCameraEvent : EffectEvent
+{
+    public float time;
+    public override IEnumerator Execute(EventObject eventObject = null)
+    {
+        return EffectManager.Instance.ShakeCamera(time);
+    }
+}
+public class ParticleEvent : EffectEvent
 {
 
     public override IEnumerator Execute(EventObject eventObject = null)
@@ -11,5 +36,23 @@ public class ParticleEvent : Event
         particleSystem.Play();
         yield return new WaitForSeconds(0.1f);
         particleSystem.Stop();
+    }
+}
+
+[Serializable]
+public class FadeInEvent : EffectEvent
+{
+    public override IEnumerator Execute(EventObject eventObject = null)
+    {
+        return EffectManager.Instance.FadeIn();
+    }
+}
+
+[Serializable]
+public class FadeOutEvent : EffectEvent
+{
+    public override IEnumerator Execute(EventObject eventObject = null)
+    {
+        return EffectManager.Instance.FadeOut();
     }
 }
